More refactoring of Date/Time tests
