+++ tests/PhpSpreadsheetTests/Calculation/Functions/DateTime/NetworkDaysTest.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\DateTime;
+
+use PhpOffice\PhpSpreadsheet\Calculation\DateTime;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Shared\Date;
+use PHPUnit\Framework\TestCase;
+
+class NetworkDaysTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        Functions::setReturnDateType(Functions::RETURNDATE_EXCEL);
+        Date::setExcelCalendar(Date::CALENDAR_WINDOWS_1900);
+    }
+
+    /**
+     * @dataProvider providerNETWORKDAYS
+     *
+     * @param mixed $expectedResult
+     */
+    public function testNETWORKDAYS($expectedResult, ...$args)
+    {
+        $result = DateTime::NETWORKDAYS(...$args);
+        $this->assertEquals($expectedResult, $result, '', 1E-8);
+    }
+
+    public function providerNETWORKDAYS()
+    {
+        return require 'data/Calculation/DateTime/NETWORKDAYS.php';
+    }
+}
